refactor: app 컴포넌트 분리
